+++ src/hcl_model/model_base.py
@@ -1,5 +1,5 @@
 from abc import ABC, abstractmethod
-from typing import List, Optional
+from typing import List, Optional, Union
 
 import numpy as np
 import pandas as pd
@@ -22,7 +22,7 @@ class ModelBase(ABC):
         self._y_train = None
         self._x_train = None
 
-    def fit(self, X: Optional[pd.DataFrame], y: pd.Series, **kwargs):
+    def fit(self, X: Optional[pd.DataFrame], y: Union[pd.Series, np.ndarray], **kwargs):
         """
         Fit the model using some provided training data.
 
@@ -40,8 +40,8 @@ class ModelBase(ABC):
 
     def predict(
         self,
+        X: Optional[pd.DataFrame],
         num_steps: int,
-        X: pd.DataFrame = None,
         quantile_levels: List[float] = None,
         num_simulations: int = None,
         **kwargs
@@ -49,8 +49,8 @@ class ModelBase(ABC):
         """
         Forecast the values and prediction intervals
 
-        :param num_steps: number of point in the future that we want to forecast
         :param X: exogenous variables should cover the whole prediction horizon
+        :param num_steps: number of point in the future that we want to forecast
         :param quantile_levels: list of desired prediction interval levels between 0 and 100 (in percentages).
             If not provided, no confidence interval will be given as output
         :param num_simulations: number of simulations for simulation-based prediction intervals