+++ src/hcl_model/utils/check_x_y.py
@@ -6,5 +6,13 @@ import pandas as pd
 
 def check_X_y(X: Optional[pd.DataFrame], y: Union[pd.Series, np.ndarray]) -> Tuple[Optional[pd.DataFrame], pd.Series]:
     X = X.copy() if X is not None else None
-    y = pd.Series(y, index=X.index, name="value") if isinstance(y, np.ndarray) else y.copy()
+    if isinstance(y, np.ndarray) & (X is not None):
+        y = pd.Series(y, index=X.index, name="value")
+    elif isinstance(y, np.ndarray):
+        raise TypeError(
+            "Passing ndarray for y and None for X is not supported. "
+            "Please, consider converting y to Series with DatetimeIndex."
+        )
+    else:
+        y = y.copy()
     return X, y